CREATE_FS: Due to the fact that Equallogic storage doesn't seem to zero newly created luns, it is possible that a newly created lun contains the beginnings of an old filesystem that was previously stored on the physical area that the new LUN uses. This would lead to problems when booting, because the old code assumed the lun did not contain anything that could be recognised as a filesystem. To solve this, we add a new deployment parameter called CREATE_FS and check if its value is true in the root-mount script. Only if it is true, a new filesystem will be created. Afterwards the command create_fs_finished is posted to the root-mount.equallogic.action.php hook in boot-services, which sets the parameter to false.
RESIZE_FS: The second part of this patch implements the same logic for resizing in the cloud. When the cloud monitor executes a resize request, the deployment parameter RESIZE_FS will be set. When this is set, the root-mount script will try to resize the image at boottime. Afterwards the command resize_fs_finished is posted to set the parameter to false.


git-svn-id: https://openqrm.svn.sourceforge.net/svnroot/openqrm/trunk@1950 f7cd0d49-994a-0410-81fc-ecfcdc9cdbae